// my.replika.ai/static/js/types/enums.ts
export enum ActionTypes {
  InitPersist = 'INIT_PERSIST',
  UpdatePersist = 'UPDATE_PERSIST',
  CheckDeviceId = 'CHECK_DEVICE_ID',
  SignupFieldChange = 'SIGNUP_FIELD_CHANGE',
  LoginFieldChange = 'LOGIN_FIELD_CHANGE',
  SetCodeResendTimeout = 'SET_CODE_RESEND_TIMEOUT',
  WsStart = 'WS_START',
  WsOpen = 'WS_OPEN',
  WsClose = 'WS_CLOSE',
  WsDisconnect = 'WS_DISCONNECT',
  WsMessageReceived = 'WS_MESSAGE_RECEIVED',
  WsError = 'WS_ERROR',
  WsChatMessageReceived = 'WS_CHAT_MESSAGE_RECEIVED',
  WsMessageReactionReceived = 'WS_MESSAGE_REACTION_RECEIVED',
  WsFeedbackRequestReceived = 'WS_FEEDBACK_REQUEST_RECEIVED',
  WsStartTypingReceived = 'WS_START_TYPING_RECEIVED',
  WsStopTypingReceived = 'WS_STOP_TYPING_RECEIVED',
  WsVoiceModeReceived = 'WS_VOICE_MODE_RECEIVED',
  WsUserMoodReceived = 'WS_USER_MOOD_RECEIVED',
  WsBotMoodReceived = 'WS_BOT_MOOD_RECEIVED',
  CloseFeedbackPopup = 'CLOSE_FEEDBACK_POPUP',
  BotStatsReceived = 'BOT_STATS_UPDATE_RECEIVED',
  BotChatReceived = 'BOT_CHAT_UPDATE_RECEIVED',
  BotReceived = 'BOT_UPDATE_RECEIVED',
  SetDefaultChat = 'SET_DEFAULT_CHAT',
  GoneOffline = 'GONE_OFFLINE',
  GoneOnline = 'GONE_ONLINE',
  SetActiveWindow = 'SET_ACTIVE_WINDOW',
  SetTheme = 'SET_THEME',
  SetTrialPopupActive = 'SET_TRIAL_POPUP_ACTIVE',
  UserpicUploadLocallySuccess = 'USERPIC_UPLOAD_LOCALLY_SUCCESS',
  UserpicUploadLocallyError = 'USERPIC_UPLOAD_LOCALLY_ERROR',
  TrialPeriodExpired = 'TRIAL_PERIOD_EXPIRED',
  ToggleMobileProfile = 'TOGGLE_MOBILE_PROFILE',
  ResetServerError = 'RESET_SERVER_ERROR',
  AcceptGdpr = 'ACCEPT_GDPR',
  GoToReboardingStep = 'GO_TO_REBOARDING_STEP',
  UpdateReboardingState = 'UPDATE_REBOARDING_STATE',
  GoToSignupStep = 'GO_TO_SIGNUP_STEP',
  UpdateSignupState = 'UPDATE_SIGNUP_STATE',
  SetHintStatus = 'SET_HINT_STATUS',
  SetHintMessageReaction = 'SET_HINT_MESSAGE_REACTION',
  UpdateChatMessageText = 'UPDATE_CHAT_MESSAGE_TEXT',
  SkipWidget = 'SKIP_WIDGET',
  ExperimentStarted = 'EXPERIMENT_STARTED',
  ExperimentEnded = 'EXPERIMENT_ENDED',
  InitNewUser = 'INIT_NEW_USER',
  UploadChatImagePreview = 'UPLOAD_CHAT_IMAGE_PREVIEW',
  SetActiveDialog = 'SET_ACTIVE_DIALOG',
  WsJourneyChanged = 'WS_JOURNEY_CHANGED',
  SetCompletedMissionDialogShown = 'SET_COMPLETED_MISSION_DIALOG_SHOWN',
  SetIntroStep = 'SET_INTRO_STEP',
  SetReboardingDialogShown = 'SET_REBOARDING_DIALOG_SHOWN',
  SetShowRelationshipStatusButton = 'SET_SHOW_RELATIONSHIP_STATUS_BUTTON',
  DismissNotification = 'DISMISS_NOTIFICATION',
  SetDisabled3d = 'SET_DISABLED_3D',
  SetChooseAvatarScreenShown = 'SET_CHOOSE_AVATAR_SCREEN_SHOWN',
  SetShowMessageActions = 'SET_SHOW_MESSAGE_ACTIONS',
  SetNewThought = 'SET_NEW_THOUGHT',
  WsSaveChatItemReceived = 'WS_SAVE_CHAT_ITEM_RECEIVED',
  CancelSaveChatItem = 'CANCEL_SAVE_CHAT_ITEM',
  UpdateChatItemToSave = 'UPDATE_CHAT_ITEM_TO_SAVE',
  WsStatementRemembered = 'WS_STATEMENT_REMEMBERED',
  ResetMemoryUnreadCount = 'RESET_MEMORY_UNREAD_COUNT',
  ClearChatMessageAlert = 'CLEAR_CHAT_MESSAGE_ALERT',
  WsNavigationPopup = 'WS_NAVIGATION_POPUP',
  SetStripeLatestInvoiceId = 'SET_STRIPE_LATEST_INVOICE_ID',
  SetStripeEnabled = 'SET_STRIPE_ENABLED',

  // required for 3rg party libs
  OtherAction = '__any_other_action_type__',
}
// ...
export enum WsEventNames {
  Error = 'error',
  Init = 'init',
  NewInit = 'new_init',
  History = 'history',
  Message = 'message',
  MessageReaction = 'message_reaction',
  Bot = 'bot',
  PersonalBotStats = 'personal_bot_stats',
  PersonalBotChat = 'personal_bot_chat',
  ConversationFeedback = 'conversation_feedback',
  StartTyping = 'start_typing',
  StopTyping = 'stop_typing',
  TextInputDetected = 'text_input_detected',
  VoiceMode = 'voice_mode',
  UserMood = 'user_mood',
  BotMood = 'bot_mood',
  AppForeground = 'app_foreground',
  ExperimentStarted = 'experiment_started',
  ExperimentEnded = 'experiment_ended',
  StartMission = 'start_mission',
  SkipMission = 'skip_mission',
  DropMission = 'drop_mission',
  JourneyChanged = 'journey_changed',
  MainScreen = 'main_screen',
  ChatScreen = 'chat_screen',
  ApplicationStarted = 'application_started',
  ReplikaThoughtViewed = 'replika_thought_viewed',
  SaveChatItemRequest = 'save_chat_item_request',
  SaveChatItem = 'save_chat_item',
  StatementRemembered = 'statement_remembered',
  NavigationPopup = 'navigation_popup',
}

export enum MetricsEvents {
  Launched = 'Application launched',
  Upvote = 'Upvote',
  Downvote = 'Downvote',
  AccountDeleted = 'Account deleted',
  NeedHelpButtonClicked = 'Need help - Button clicked',
  NeedHelpOptionSelected = 'Need help - Option selected',
  NeedHelpHotlineCalled = 'Need help - Hotline called',
  NeedHelpLifelineChatClicked = 'Need help - Lifeline chat clicked',
  NeedHelpWebsiteClicked = 'Need help - Website clicked',
  ExperimentStarted = 'Experiment Started',
  ExperimentEnded = 'Experiment Ended',
  ForgotPassword = 'Forgot password',
  MoodAdded = 'Mood added',
  MoodEdited = 'Mood edited',
  MoodDeleted = 'Mood deleted',
  MoodSliderCalled = 'Mood slider called',
  MoodOpened = 'Mood opened',
  MoodSliderDisplayed = 'Mood slider displayed',
  MoodSliderSkipped = 'Mood slider skipped',
  WelcomeScreenOpened = 'Welcome screen opened',
  SignupScreenOpened = 'Signup screen opened',
  LoginScreenOpened = 'Login screen opened',
  TrackOpened = 'Track opened',
  CategoryOpened = 'Category opened',
  SpotlightOpened = 'Spotlight opened',
  ReplikaImageSet = 'Replika image set',
  CreditsOpened = 'Credits opened',
  HelpCenterOpened = 'Help center opened',
  SettingsOpened = 'Settings opened',
  PrivacyPolicyOpened = 'Privacy policy opened',
  TermsOfServiceOpend = 'Terms of service opened',
  PhraseShown = 'Phrase shown',
  UserAccountSet = 'User account set',
  ReplikaProfileSet = 'Replika profile set',
  ReplikaNameAdded = 'Replika name added',
  TestimonialShown = 'Testimonial shown',
  FirstOnboardingPhraseShown = '1st onboarding phrase shown',
  MeetButtonShown = 'Meet button shown',
  MeetButtonPressed = 'Meet button pressed',
  RelationshipStatusOpened = 'Relationship status opened',
  RelationshipStatusClosed = 'Relationship status closed',
  RelationshipStatusChanged = 'Relationship status changed',
  RelationshipStatusAttemptToChange = 'Relationship status attempt to change',
  SubscriptionWebPromptShown = 'Subscription web prompt shown',
  SubscriptionWebPromptClicked = 'Subscription web prompt clicked',
  SubscriptionExtendedTrialShown = 'Subscription extended trial shown',
  PageIsAutoTranslated = 'Page is auto-translated',
  SavedChatItemsPopupOpened = 'Save chat elements: pop-up opened',
  SavedChatItemsSaveButtonPressed = 'Save chat elements: save button pressed',
  SavedChatItemsCancelButtonPressed = 'Save chat elements: cancel button pressed',
  SavedChatItemsPopupEditNamePressed = 'Save chat elements: Pop-up edit name pressed',
  SavedChatItemsLibraryEditNamePressed = 'Save chat elements: Library edit name pressed',
  SavedChatItemOpened = 'Save chat elements: Element opened',
  SavedChatItemsEmptyStateButtonPressed = 'Save chat elements: Empty state pressed',
}
// ...
export enum Dialogs {
  Feedback = 'Feedback',
  EditProfile = 'EditProfile',
  EditReplika = 'EditReplika',
  AccountSettings = 'AccountSettings',
  EditPassword = 'EditPassword',
  DropMission = 'DropMission',
  Badge = 'Badge',
  EditMood = 'EditMood',
  IntroMeetReplika = 'IntroMeetReplika',
  GetHelp = 'GetHelp',
  SubscriptionReboarding = 'SubscriptionReboarding',
  Subscription = 'Subscription',
  EditRelationshipStatus = 'EditRelationshipStatus',
  ReboardingAddPronoun = 'ReboardingAddPronoun',
  RewardPopup = 'RewardPopup',
  CustomizeAvatar = 'CustomizeAvatar',
  Reaction = 'Reaction',
  EditSavedChatItem = 'EditSavedChatItem',
  NavigationPopup = 'NavigationPopup',
  CancelSubscription = 'CancelSubscription',
}